+++ src/shared/lib/theme/ThemeProvider.tsx
@@ -1,6 +1,6 @@
 import type { FC, ReactNode } from "react";
 import { useState } from "react";
-import ThemeContext from "@/shared/lib/theme/themeContext"
+import ThemeContext from "@/shared/lib/theme/ThemeContext"
 
 type ThemeProviderProps = {
 	children: ReactNode;
