+++ src/shared/lib/theme/useTheme.ts
@@ -1,11 +1,11 @@
 import { useContext } from "react";
-import ThemeContext from "@/shared/lib/theme/themeContext";
+import ThemeContext from "@/shared/lib/theme/ThemeContext";
 
 export const useTheme = () => {
 	const context = useContext(ThemeContext);
 
 	if (context === undefined) {
-		console.log("useThemeContext error");
+		throw new Error("useContext === undefined");
 	}
 
 	return context;
